Fix cublas build issue
Fix "no such file or directory" do_compile error step.

It seems the /usr/lib/aarch64-linux-gnu/ directory does not
exist  when running this recipe, so add support for skipping
relevant steps.

+++ recipes-devtools/cuda/libcublas_10.2.2.89-1.bb
@@ -11,9 +11,11 @@ SRC_URI[main.sha256sum] = "d0299b139a163136432dfb2c028769944b6c5636ad9238614860c
 SRC_URI[dev.sha256sum] = "5fa7e3e8fe266fdea7e91778610b7e8d3d85d8950875a4915ce3626c9e564365"
 
 do_compile_append() {
-    mv ${B}/usr/${baselib}/${HOST_ARCH}-linux-gnu/lib* ${B}/usr/${baselib}/
-    rm -rf ${B}/usr/${baselib}/${HOST_ARCH}-linux-gnu/stubs
-    rmdir ${B}/usr/${baselib}/${HOST_ARCH}-linux-gnu
+    if [ -d ${B}/usr/${baselib}/${HOST_ARCH}-linux-gnu/ ]; then
+	    mv ${B}/usr/${baselib}/${HOST_ARCH}-linux-gnu/lib* ${B}/usr/${baselib}/
+	    rm -rf ${B}/usr/${baselib}/${HOST_ARCH}-linux-gnu/stubs
+	    rmdir ${B}/usr/${baselib}/${HOST_ARCH}-linux-gnu
+    fi
     sed -i -e's,^pkgroot=.*,prefix=${prefix},' -e's,{pkgroot},{prefix},g' \
 	-e's! -Wl,-rpath.*!!' ${B}/usr/${baselib}/pkgconfig/cublas-10.pc
     ln -s cublas-10.pc ${B}/usr/${baselib}/pkgconfig/cublas-10.2.pc
